requirements.txt
libvirt-python
flask
flask_cors
uwsgi
typer